1 hr
Made a cool noisy rainbow and the mandelbrot set. The rainbow was pretty easy and fast, the mandelbrot set was cool to make. I still don't fully get it, but I'm working on the explanation right now to try to help me get my own head to wrap around what I made.